refactor(frontend): decompose TeacherDashboard into section components
TeacherDashboard.tsx shrank from 529 lines to 284 lines. The page now
orchestrates data loading + CRUD actions and delegates each visual
section to a focused component under `pages/Teacher/dashboard/`:

- CreateCourseForm.tsx  — inline create-course form card
- PendingCertsCard.tsx  — pending-certificate approve/reject card
- CourseCard.tsx        — single course row with full action bar
- EmptyCoursesCard.tsx  — empty state CTA
- TrashSection.tsx      — self-contained trash section with its own
                          load/restore/permanent-delete lifecycle
- types.ts              — `PendingCert` alias
- index.ts              — barrel exports

`TrashSection` now owns its own fetch + state, removing the equivalent
logic from the parent. The parent only needs to know about a single
`onRestore(Course)` callback to re-inject restored courses back into
the main list.

No functional/behavioral changes. tsc, eslint, vitest (85/85) and
`npm run build` all pass.

Made-with: Cursor
