+++ frontend/src/pages/Teacher/dashboard/index.ts
@@ -1,0 +1,6 @@
+export { CourseCard } from "./CourseCard"
+export { CreateCourseForm } from "./CreateCourseForm"
+export { EmptyCoursesCard } from "./EmptyCoursesCard"
+export { PendingCertsCard } from "./PendingCertsCard"
+export { TrashSection } from "./TrashSection"
+export type { PendingCert } from "./types"
