+++ frontend/src/pages/Teacher/dashboard/types.ts
@@ -1,0 +1,6 @@
+import type { Certificate } from "@/types"
+
+export type PendingCert = Certificate & {
+  student_name?: string
+  course_title?: string
+}
